+++ lib/src/youtube_player.dart
@@ -11,6 +11,7 @@ bool _justSwitchedToFullScreen = false;
 
 /// Current state of the player. Find more about it [here](https://developers.google.com/youtube/iframe_api_reference#Playback_status)
 enum PlayerState {
+  UNKNOWN,
   UN_STARTED,
   ENDED,
   PLAYING,
@@ -51,6 +52,10 @@ class YoutubePlayer extends StatefulWidget {
   /// Default = true
   final bool autoPlay;
 
+  /// Mutes the player initially
+  /// Default = false
+  final bool mute;
+
   /// Overrides the default buffering indicator for the player.
   final Widget bufferIndicator;
 
@@ -74,6 +79,18 @@ class YoutubePlayer extends StatefulWidget {
   /// Overrides color of Live UI when enabled.
   final Color liveUIColor;
 
+  /// Hides the fullscreen button.
+  final bool hideFullScreenButton;
+
+  /// Adds custom top bar widgets
+  final List<Widget> actions;
+
+  /// If true, hides the YouTube player annotation. Default = false
+  ///
+  /// Forcing annotation to hide is a hacky way. Although this shouldn't be against Youtube TOS, the author doesn't guarantee
+  /// and won't be responsible for any casualties regarding the YouTube TOS violation.
+  final bool forceHideAnnotation;
+
   YoutubePlayer({
     Key key,
     @required this.context,
@@ -81,6 +98,7 @@ class YoutubePlayer extends StatefulWidget {
     this.width,
     this.aspectRatio = 16 / 9,
     this.autoPlay = true,
+    this.mute = false,
     this.hideControls = false,
     this.controlsTimeOut = const Duration(seconds: 3),
     this.bufferIndicator,
@@ -91,6 +109,9 @@ class YoutubePlayer extends StatefulWidget {
     this.isLive = false,
     this.liveUIColor = Colors.red,
     this.thumbnail,
+    this.hideFullScreenButton = false,
+    this.actions,
+    this.forceHideAnnotation = false,
   })  : assert(videoId.length == 11, "Invalid YouTube Video Id"),
         super(key: key);
 
@@ -153,9 +174,6 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
 
   _loadController({WebViewController webController}) {
     controller = YoutubePlayerController(widget.videoId);
-    if (webController != null)
-      controller.value =
-          controller.value.copyWith(webViewController: webController);
     controller.addListener(listener);
     if (widget.onPlayerInitialized != null)
       widget.onPlayerInitialized(controller);
@@ -188,8 +206,9 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
       _isFullScreen = false;
       Future.delayed(
         Duration(milliseconds: 500),
-        () => controller
-            .seekTo(Duration(milliseconds: _oldPosition.inMilliseconds + 500)),
+        () => controller.seekTo(
+              Duration(milliseconds: _oldPosition.inMilliseconds + 500),
+            ),
       );
     }
   }
@@ -204,7 +223,7 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
   Widget build(BuildContext context) {
     if (_currentVideoId != widget.videoId) {
       _currentVideoId = widget.videoId;
-      _loadController(webController: controller.value.webViewController);
+      _loadController();
       controller.load();
       Future.delayed(Duration(milliseconds: 500),
           () => controller.seekTo(Duration(seconds: 0)));
@@ -225,6 +244,8 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
           _Player(
             controller: controller,
             autoPlay: widget.autoPlay,
+            forceHideAnnotation: widget.forceHideAnnotation,
+            mute: widget.mute,
           ),
           (controller.value.isPlaying ||
                       controller.value.playerState == PlayerState.PAUSED) ||
@@ -243,78 +264,85 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
                     fit: BoxFit.cover,
                   ),
                 ),
-          widget.hideControls
-              ? Container()
-              : TouchShutter(
-                  controller,
-                  _showControls,
-                ),
-          widget.hideControls
-              ? Container()
-              : controller.value.position > Duration(milliseconds: 100) &&
-                      !_showControls.value &&
-                      widget.showVideoProgressIndicator &&
-                      !widget.isLive &&
-                      !controller.value.isFullScreen
-                  ? Positioned(
-                      bottom: -27.9,
-                      left: 0,
-                      right: 0,
-                      child: IgnorePointer(
-                        ignoring: true,
-                        child: ProgressBar(
-                          controller,
-                          colors: ProgressColors(
-                            handleColor: Colors.transparent,
-                            playedColor: widget.videoProgressIndicatorColor,
-                          ),
+          if (!widget.hideControls)
+            TouchShutter(
+              controller,
+              _showControls,
+            ),
+          if (!widget.hideControls)
+            (controller.value.position > Duration(milliseconds: 100) &&
+                    !_showControls.value &&
+                    widget.showVideoProgressIndicator &&
+                    !widget.isLive &&
+                    !controller.value.isFullScreen)
+                ? Positioned(
+                    bottom: -27.9,
+                    left: 0,
+                    right: 0,
+                    child: IgnorePointer(
+                      ignoring: true,
+                      child: ProgressBar(
+                        controller,
+                        colors: ProgressColors(
+                          handleColor: Colors.transparent,
+                          playedColor: widget.videoProgressIndicatorColor,
                         ),
                       ),
+                    ),
+                  )
+                : Container(),
+          if (!widget.hideControls)
+            Positioned(
+              bottom: 0,
+              left: 0,
+              right: 0,
+              child: widget.isLive
+                  ? LiveBottomBar(
+                      controller,
+                      _showControls,
+                      widget.aspectRatio,
+                      widget.liveUIColor,
+                      widget.hideFullScreenButton,
                     )
-                  : Container(),
-          widget.hideControls
-              ? Container()
-              : Positioned(
-                  bottom: 0,
-                  left: 0,
-                  right: 0,
-                  child: widget.isLive
-                      ? LiveBottomBar(
-                          controller,
-                          _showControls,
-                          widget.aspectRatio,
-                          widget.liveUIColor,
-                        )
-                      : BottomBar(
-                          controller,
-                          _showControls,
-                          widget.aspectRatio,
-                          widget.progressColors,
-                        ),
-                ),
-          widget.hideControls
-              ? Container()
-              : Center(
-                  child: PlayPauseButton(
-                    controller,
-                    _showControls,
-                    widget.bufferIndicator ??
-                        Container(
-                          width: 70.0,
-                          height: 70.0,
-                          child: CircularProgressIndicator(
-                            valueColor: AlwaysStoppedAnimation(Colors.white),
-                          ),
-                        ),
-                  ),
-                ),
+                  : BottomBar(
+                      controller,
+                      _showControls,
+                      widget.aspectRatio,
+                      widget.progressColors,
+                      widget.hideFullScreenButton,
+                    ),
+            ),
+          if (!widget.hideControls && _showControls.value)
+            Positioned(
+              top: 0,
+              left: 0,
+              right: 0,
+              child: Row(
+                children: widget.actions ?? [Container()],
+              ),
+            ),
+          if (!widget.hideControls)
+            Center(
+              child: PlayPauseButton(
+                controller,
+                _showControls,
+                widget.bufferIndicator ??
+                    Container(
+                      width: 70.0,
+                      height: 70.0,
+                      child: CircularProgressIndicator(
+                        valueColor: AlwaysStoppedAnimation(Colors.white),
+                      ),
+                    ),
+              ),
+            ),
         ],
       ),
     );
   }
 
   Future<void> _pushFullScreenWidget(BuildContext context) async {
-    _oldWebController = controller.value.webViewController;
+    _oldWebController = controller._currentWebController;
     _justSwitchedToFullScreen = false;
     _oldPosition = controller.value.position;
     controller.pause();
@@ -349,8 +377,8 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
     String popValue = await Navigator.of(context).push(route);
     if (popValue == null) _isFullScreen = false;
 
-    controller.value = controller.value
-        .copyWith(isFullScreen: false, webViewController: _oldWebController);
+    controller.value = controller.value.copyWith(isFullScreen: false);
+    controller._currentWebController = _oldWebController;
 
     Future.delayed(
       Duration(milliseconds: 500),
@@ -370,8 +398,11 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
     if (popValue == null)
       Future.delayed(
         Duration(milliseconds: 500),
-        () => controller
-            .seekTo(Duration(milliseconds: _oldPosition.inMilliseconds + 500)),
+        () => controller.seekTo(
+              Duration(
+                milliseconds: _oldPosition.inMilliseconds + 500,
+              ),
+            ),
       );
   }
 }
@@ -379,23 +410,30 @@ class _YoutubePlayerState extends State<YoutubePlayer> {
 class _Player extends StatefulWidget {
   final YoutubePlayerController controller;
   final bool autoPlay;
+  final bool forceHideAnnotation;
+  final bool mute;
 
   _Player({
     this.controller,
     this.autoPlay,
+    this.forceHideAnnotation,
+    this.mute,
   });
 
   @override
   __PlayerState createState() => __PlayerState();
 }
 
-class __PlayerState extends State<_Player> with AutomaticKeepAliveClientMixin {
+class __PlayerState extends State<_Player> {
+  WebViewController _temp;
+
   @override
   Widget build(BuildContext context) {
     return IgnorePointer(
       ignoring: true,
       child: WebView(
-        initialUrl: "https://sarbagyadhaubanjar.github.io/youtube_player",
+        initialUrl:
+            "https://sarbagyadhaubanjar.github.io/youtube_player/youtube.html",
         javascriptMode: JavascriptMode.unrestricted,
         iOSWebViewConfiguration: IOSWebViewConfiguration(
           allowsInlineMediaPlayback: true,
@@ -467,7 +505,7 @@ class __PlayerState extends State<_Player> with AutomaticKeepAliveClientMixin {
             name: 'Errors',
             onMessageReceived: (JavascriptMessage message) {
               widget.controller.value = widget.controller.value
-                  .copyWith(errorCode: int.parse(message.message ?? 0));
+                  .copyWith(errorCode: int.tryParse(message.message) ?? 0);
             },
           ),
           JavascriptChannel(
@@ -486,12 +524,7 @@ class __PlayerState extends State<_Player> with AutomaticKeepAliveClientMixin {
           JavascriptChannel(
             name: 'CurrentTime',
             onMessageReceived: (JavascriptMessage message) {
-              double position;
-              try {
-                position = double.parse(message.message) * 1000;
-              } catch (e) {
-                position = 0;
-              }
+              double position = (double.tryParse(message.message) ?? 0) * 1000;
               widget.controller.value = widget.controller.value.copyWith(
                 position: Duration(
                   milliseconds: position.floor(),
@@ -509,17 +542,23 @@ class __PlayerState extends State<_Player> with AutomaticKeepAliveClientMixin {
                 positionLoadedFraction = 0;
               }
               widget.controller.value = widget.controller.value.copyWith(
-                buffered: positionLoadedFraction,
+                buffered: double.tryParse(message.message) ?? 0,
               );
             },
           ),
         ].toSet(),
         onWebViewCreated: (webController) {
-          widget.controller.value = widget.controller.value
-              .copyWith(webViewController: webController);
+          _temp = webController;
         },
         onPageFinished: (_) {
-          print("finished");
+          if (widget.controller.value.isFullScreen) {
+            widget.controller._currentWebController = _temp;
+          } else {
+            widget.controller.value.webViewController.complete(_temp);
+          }
+          if (widget.forceHideAnnotation)
+            widget.controller.forceHideAnnotation();
+          if (widget.mute) widget.controller.mute();
           if (widget.controller.value.isReady)
             widget.autoPlay
                 ? widget.controller.load()
@@ -528,9 +567,6 @@ class __PlayerState extends State<_Player> with AutomaticKeepAliveClientMixin {
       ),
     );
   }
-
-  @override
-  bool get wantKeepAlive => true;
 }
 
 /// [ValueNotifier] for [YoutubePlayerController].
@@ -545,10 +581,9 @@ class YoutubePlayerValue {
     this.isPlaying = false,
     this.isFullScreen = false,
     this.volume = 100,
-    this.playerState,
-    this.errorCode,
-    this.webViewController,
-  });
+    this.playerState = PlayerState.UNKNOWN,
+    this.errorCode = 0,
+  }) : webViewController = Completer<WebViewController>();
 
   /// This is true when underlying web player reports ready.
   final bool isReady;
@@ -585,7 +620,7 @@ class YoutubePlayerValue {
   final int errorCode;
 
   /// Reports the [WebViewController].
-  final WebViewController webViewController;
+  final Completer<WebViewController> webViewController;
 
   /// Returns true is player has errors.
   bool get hasError => errorCode != 0;
@@ -602,7 +637,6 @@ class YoutubePlayerValue {
     double volume,
     PlayerState playerState,
     int errorCode,
-    WebViewController webViewController,
   }) {
     return YoutubePlayerValue(
       isReady: isReady ?? this.isReady,
@@ -616,7 +650,6 @@ class YoutubePlayerValue {
       volume: volume ?? this.volume,
       playerState: playerState ?? this.playerState,
       errorCode: errorCode ?? this.errorCode,
-      webViewController: webViewController ?? this.webViewController,
     );
   }
 
@@ -637,38 +670,51 @@ class YoutubePlayerValue {
 
 class YoutubePlayerController extends ValueNotifier<YoutubePlayerValue> {
   final String initialSource;
+  WebViewController _currentWebController;
 
   YoutubePlayerController([
     this.initialSource = "",
   ]) : super(YoutubePlayerValue(isReady: false));
 
+  _evaluateJS(String javascriptString) {
+    if (_currentWebController != null) {
+      _currentWebController.evaluateJavascript(javascriptString);
+    }
+    value.webViewController.future.then(
+      (controller) {
+        _currentWebController = controller;
+        controller.evaluateJavascript(javascriptString);
+      },
+    );
+  }
+
+  /// Hide YouTube Player annotations like title, share button and youtube logo.
+  void forceHideAnnotation() => _evaluateJS('hideAnnotations()');
+
   /// Plays the video.
-  void play() =>
-      value.webViewController.evaluateJavascript('player.playVideo()');
+  void play() => _evaluateJS('play()');
 
   /// Pauses the video.
-  void pause() =>
-      value.webViewController.evaluateJavascript('player.pauseVideo()');
+  void pause() => _evaluateJS('pause()');
 
   /// Loads the video as per the [videoId] provided.
-  void load({int startAt = 0}) => value.webViewController
-      .evaluateJavascript('player.loadVideoById("$initialSource", $startAt)');
+  void load({int startAt = 0}) =>
+      _evaluateJS('loadById("$initialSource", $startAt)');
 
   /// Cues the video as per the [videoId] provided.
-  void cue({int startAt = 0}) => value.webViewController
-      .evaluateJavascript('player.cueVideoById("$initialSource", $startAt)');
+  void cue({int startAt = 0}) =>
+      _evaluateJS('cueById("$initialSource", $startAt)');
 
   /// Mutes the player.
-  void mute() => value.webViewController.evaluateJavascript('player.mute()');
+  void mute() => _evaluateJS('mute()');
 
   /// Un mutes the player.
-  void unMute() =>
-      value.webViewController.evaluateJavascript('player.unMute()');
+  void unMute() => _evaluateJS('unMute()');
 
   /// Sets the volume of player.
   /// Max = 100 , Min = 0
   void setVolume(int volume) => volume >= 0 && volume <= 100
-      ? value.webViewController.evaluateJavascript('player.setVolume($volume)')
+      ? _evaluateJS('setVolume($volume)')
       : throw Exception("Volume should be between 0 and 100");
 
   /// Seek to any position. Video auto plays after seeking.
@@ -676,8 +722,7 @@ class YoutubePlayerController extends ValueNotifier<YoutubePlayerValue> {
   /// if the seconds parameter specifies a time outside of the currently buffered video data.
   /// Default allowSeekAhead = true
   void seekTo(Duration position, {bool allowSeekAhead = true}) {
-    value.webViewController.evaluateJavascript(
-        'player.seekTo(${position.inSeconds},$allowSeekAhead)');
+    _evaluateJS('seekTo(${position.inSeconds},$allowSeekAhead)');
     play();
     value = value.copyWith(position: position);
   }