+++ CHANGELOG.md
@@ -1,3 +1,22 @@
+## 2.0.0
+* **FIXED** iOS playback is fully functional. [Issue #2](https://github.com/sarbagyastha/youtube_player_flutter/issues/2)
+* Added `forceHideAnnotation` property which hides the default YouTube annotation. [Read Issue #17 for detail](https://github.com/sarbagyastha/youtube_player_flutter/issues/14)
+* Added option to `mute` player.
+
+## 1.2.0
+* Added `actions` property, which can be used to add menus in top bar of video. See example for details.
+* Added option to hide fullscreen button. 
+* **FIXED** Current position always showing 00:00 [Issue #17](https://github.com/sarbagyastha/youtube_player_flutter/issues/17)
+
+## 1.1.1+1
+* Updated dart constraint to `>=2.2.0 <3.0.0`*. 
+* **FIXED** Parsing Exceptions.
+
+## 1.1.1
+* Fix for issue
+  [#12](https://github.com/sarbagyastha/youtube_player_flutter/issues/12).
+* WebView is now exported alongside YoutubePlayer.
+
 ## 1.1.0
 * Minor Improvements.
 * Built and tested on iOS device.